requirements.txt
click~=7.1.2
tabulate~=0.8.7
requests~=2.25.1
six~=1.15.0
defusedxml~=0.6.0
pyjwt~=1.7.1
python-openstackclient

PyYAML~=5.3.1
setuptools~=51.0.0